+++ src/main/java/com/example/calculator/service/operation/Addition.java
@@ -4,7 +4,7 @@ import com.example.calculator.contract.CalculationOperationInterface;
 
 public class Addition implements CalculationOperationInterface {
   @Override
-  public float calculate(float a, float b) {
+  public double calculate(double a, double b) {
     return a+b;
   }
 }
