+++ src/main/java/com/example/calculator/service/operation/Division.java
@@ -4,7 +4,7 @@ import com.example.calculator.contract.CalculationOperationInterface;
 
 public class Division implements CalculationOperationInterface {
   @Override
-  public float calculate(float a, float b) {
+  public double calculate(double a, double b) {
     if (b == 0) {
       throw new IllegalArgumentException("Division by zero in division operation");
     }
