+++ src/main/java/com/example/calculator/service/operation/Exponentiation.java
@@ -4,11 +4,11 @@ import com.example.calculator.contract.CalculationOperationInterface;
 
 public class Exponentiation implements CalculationOperationInterface {
   @Override
-  public float calculate(float a, float b) {
+  public double calculate(double a, double b) {
     if (b < 0) {
       throw new ArithmeticException("power in negative state not allowed");
     }
 
-    return (float) Math.pow(a, b);
+    return Math.pow(a, b);
   }
 }
